refactor: split decision kernel behind stable facade
